+++ Frontend/src/de/numericat/Main.java
@@ -14,7 +14,6 @@ import java.util.Scanner;
 public class Main {
 
     private static Numericanvas canvas;
-    private static String outputPath;
     private static Timer timer;
 
     public static void main(String[] args) {
@@ -22,15 +21,15 @@ public class Main {
             System.err.println("Expected path to output file of backend");
             return;
         }
-        outputPath = args[0];
-        System.out.println("Got output file: " + outputPath);
+        Io.outputPath = args[0];
+        System.out.println("Got output file: " + Io.outputPath);
         createWindow();
 
-        final double[] scalingFactor =  getScalingFactor(outputPath);
+        final double[] scalingFactor =  Io.getScalingFactor(canvas.getHeight());
         ActionListener action = e -> {
-            List<List<Coordinate>> data = getData(scalingFactor);
+            List<List<Coordinate>> data = Io.getData(scalingFactor);
             if (data == null) {
-                currentLine = 0;
+                Io.reset();
                 return;
             }
             canvas.drawMultipleData(data);
@@ -53,132 +52,5 @@ public class Main {
         frame.setPreferredSize(new Dimension(1000, 700));
         frame.pack();
         frame.setVisible(true);
-    }
-
-    private static boolean skipLines(Scanner scanner, int lines) {
-        for (int i = 0; i < lines; i++)
-            if (!scanner.hasNextLine())
-                return false;
-            else
-                scanner.nextLine();
-        return true;
-    }
-
-    private static boolean skipTillEmptyLine(Scanner scanner) {
-        while (scanner.hasNextLine())
-            if (scanner.nextLine().isEmpty())
-                return true;
-        return false;
-    }
-
-    private static int currentLine = 0;
-
-    private static List<List<Coordinate>> getData(double[] scalingFactor) {
-        String line;
-        String potentialLine;
-        FileInputStream fis = null;
-        Scanner sc = null;
-        try {
-            try {
-                fis = new FileInputStream(outputPath);
-                sc = new Scanner(fis);
-
-                if (!skipLines(sc, currentLine) || !sc.hasNextLine())
-                    return null;
-                line = sc.nextLine();
-
-                if (!skipTillEmptyLine(sc) || !skipLines(sc, currentLine) || !sc.hasNextLine())
-                    return null;
-                potentialLine = sc.nextLine();
-
-            } catch (FileNotFoundException e) {
-                e.printStackTrace();
-                return null;
-            }
-            if (line == null || Objects.equals(line, "")) {
-                System.out.println("Invalid line: " + line);
-                return null;
-            }
-
-            currentLine++;
-            List<List<Coordinate>> res = new ArrayList<>(2);
-            res.add(lineToCoordinates(line, scalingFactor[0]));
-            if (potentialLine == null) {
-                System.err.println("Didn't find potential line");
-                return res;
-            }
-            res.add(lineToCoordinates(potentialLine, scalingFactor[1]));
-            try {
-                fis.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            return res;
-        } finally {
-            try {
-                if (fis != null)
-                    fis.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            if (sc != null)
-                sc.close();
-        }
-    }
-
-    private static double getMaximumOfBlock(Scanner scanner) {
-        double maximum = 0;
-        while(scanner.hasNextLine()) {
-            String line = scanner.nextLine();
-            if (line.isEmpty())
-                break;
-            String[] splitString = line.split(" ");
-            for (int i = 0; i < splitString.length; i++) {
-                double tmp_Parse = Double.parseDouble(splitString[i]);
-                if (maximum < tmp_Parse) {
-                    maximum = tmp_Parse;
-                }
-            }
-        }
-        return maximum;
-    }
-
-    private static double[] getScalingFactor(String outputPath) {
-        double[] scalingFactor = {0.0,0.0};
-        Scanner scanner = null;
-        FileInputStream inputStream = null;
-        try {
-            inputStream = new FileInputStream(outputPath);
-            scanner = new Scanner(inputStream);
-            scalingFactor[0] = getMaximumOfBlock(scanner);
-            scalingFactor[1] = getMaximumOfBlock(scanner);
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } finally {
-            try {
-                if (inputStream != null)
-                    inputStream.close();
-                if (scanner != null)
-                    scanner.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            System.out.println("Scaling Factor 1:" + scalingFactor[0]);
-            System.out.println("Scaling Factor 2:" + scalingFactor[1]);
-        }
-        return scalingFactor;
-    }
-
-    private static List<Coordinate> lineToCoordinates(String line, double scalingFactor) {
-        String[] split = line.split(" ");
-        double el = 0;
-        double height = canvas.getHeight();
-        List<Coordinate> calculated = new ArrayList<>(split.length);
-        for (int i = 0; i < split.length; i++) {
-            el = (height / scalingFactor) * Double.parseDouble(split[i]);
-            calculated.add(new Coordinate(i + 1, el));
-        }
-        return calculated;
     }
 }
